+++ service/src/test/java/bio/terra/workspace/common/fixtures/ControlledAwsResourceFixtures.java
@@ -1,12 +1,10 @@
 package bio.terra.workspace.common.fixtures;
 
-import static bio.terra.workspace.common.fixtures.WorkspaceFixtures.DEFAULT_SPEND_PROFILE_ID;
 import static bio.terra.workspace.common.fixtures.WorkspaceFixtures.DEFAULT_USER_EMAIL;
+import static bio.terra.workspace.common.utils.AwsTestUtils.AWS_REGION;
 import static software.amazon.awssdk.services.sagemaker.model.InstanceType.ML_T2_MEDIUM;
 
-import bio.terra.aws.resource.discovery.Environment;
-import bio.terra.aws.resource.discovery.LandingZone;
-import bio.terra.aws.resource.discovery.Metadata;
+import bio.terra.workspace.common.utils.AwsTestUtils;
 import bio.terra.workspace.common.utils.AwsUtils;
 import bio.terra.workspace.common.utils.TestUtils;
 import bio.terra.workspace.generated.model.ApiAwsS3StorageFolderCreationParameters;
@@ -14,24 +12,16 @@ import bio.terra.workspace.generated.model.ApiAwsSageMakerNotebookCreationParame
 import bio.terra.workspace.service.resource.controlled.cloud.aws.s3StorageFolder.ControlledAwsS3StorageFolderResource;
 import bio.terra.workspace.service.resource.controlled.cloud.aws.sageMakerNotebook.ControlledAwsSageMakerNotebookResource;
 import bio.terra.workspace.service.resource.controlled.model.AccessScopeType;
-import bio.terra.workspace.service.resource.model.WsmResourceState;
-import bio.terra.workspace.service.workspace.model.AwsCloudContext;
-import bio.terra.workspace.service.workspace.model.AwsCloudContextFields;
-import bio.terra.workspace.service.workspace.model.AwsCloudContextFields.AwsCloudContextV1;
-import bio.terra.workspace.service.workspace.model.CloudContextCommonFields;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
-import software.amazon.awssdk.arns.Arn;
 import software.amazon.awssdk.auth.credentials.AnonymousCredentialsProvider;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
 import software.amazon.awssdk.awscore.exception.AwsServiceException;
 import software.amazon.awssdk.core.internal.waiters.DefaultWaiterResponse;
 import software.amazon.awssdk.core.waiters.WaiterResponse;
 import software.amazon.awssdk.http.SdkHttpResponse;
-import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.model.DeleteObjectsResponse;
 import software.amazon.awssdk.services.s3.model.ListObjectsV2Response;
 import software.amazon.awssdk.services.s3.model.PutObjectResponse;
@@ -45,25 +35,9 @@ import software.amazon.awssdk.services.sagemaker.model.StartNotebookInstanceResp
 import software.amazon.awssdk.services.sagemaker.model.StopNotebookInstanceResponse;
 import software.amazon.awssdk.services.sts.model.Tag;
 
+// fixtures for controlled AWS resources
 public class ControlledAwsResourceFixtures {
-  public static final long V1_VERSION = AwsCloudContextV1.getVersion();
-  public static final String MAJOR_VERSION = "v0.5.8";
-  public static final String ORGANIZATION_ID = "o-organization";
-  public static final String ACCOUNT_ID = "1245893245";
-  public static final String TENANT_ALIAS = "tenant-saas";
-  public static final String ENVIRONMENT_ALIAS = "unit-test-env";
-  public static final String AWS_REGION = "us-east-1";
-  public static final String AWS_ENVIRONMENT_WSM_ROLE_ARN =
-      "arn:aws:iam::10000000001:role/WorkspaceManagerRole";
-  public static final String AWS_ENVIRONMENT_USER_ROLE_ARN =
-      "arn:aws:iam::20000000002:role/UserRole";
-  public static final String AWS_ENVIRONMENT_NOTEBOOK_ROLE_ARN =
-      "arn:aws:iam::30000000003:role/NotebookRole";
-  public static final String AWS_LANDING_ZONE_STORAGE_BUCKET_ARN =
-      "arn:aws:iam::40000000004:role/StorageBucket";
-  public static final String AWS_LANDING_ZONE_KMS_KEY_ARN = "arn:aws:iam::50000000005:role/KmsKey";
-  public static final String AWS_LANDING_ZONE_NOTEBOOK_LIFECYCLE_CONFIG_ARN =
-      "arn:aws:iam::60000000006:role/NotebookLifecycleConfig";
+
   public static final SdkHttpResponse SDK_HTTP_RESPONSE_200 =
       SdkHttpResponse.builder().statusCode(200).statusText("Success").build();
   public static final SdkHttpResponse SDK_HTTP_RESPONSE_400 =
@@ -75,47 +49,10 @@ public class ControlledAwsResourceFixtures {
   public static final AwsServiceException AWS_SERVICE_EXCEPTION_2 =
       AwsServiceException.builder().message("not authorized to perform").build();
 
-  public static final Metadata AWS_METADATA =
-      Metadata.builder()
-          .tenantAlias(TENANT_ALIAS)
-          .organizationId(ORGANIZATION_ID)
-          .environmentAlias(ENVIRONMENT_ALIAS)
-          .accountId(ACCOUNT_ID)
-          .region(Region.of(AWS_REGION))
-          .majorVersion(MAJOR_VERSION)
-          .tagMap(Map.of("tagKey", "tagValue"))
-          .build();
-  public static final LandingZone AWS_LANDING_ZONE =
-      LandingZone.builder()
-          .metadata(AWS_METADATA)
-          .storageBucket(Arn.fromString(AWS_LANDING_ZONE_STORAGE_BUCKET_ARN), "bucket")
-          .kmsKey(Arn.fromString(AWS_LANDING_ZONE_KMS_KEY_ARN), UUID.randomUUID())
-          .addNotebookLifecycleConfiguration(
-              Arn.fromString(AWS_LANDING_ZONE_NOTEBOOK_LIFECYCLE_CONFIG_ARN), "lifecycleConfig")
-          .build();
-  public static final Environment AWS_ENVIRONMENT =
-      Environment.builder()
-          .metadata(AWS_METADATA)
-          .workspaceManagerRoleArn(Arn.fromString(AWS_ENVIRONMENT_WSM_ROLE_ARN))
-          .userRoleArn(Arn.fromString(AWS_ENVIRONMENT_USER_ROLE_ARN))
-          .notebookRoleArn(Arn.fromString(AWS_ENVIRONMENT_NOTEBOOK_ROLE_ARN))
-          .addLandingZone(AWS_LANDING_ZONE.getMetadata().getRegion(), AWS_LANDING_ZONE)
-          .build();
-
   public static Collection<Tag> makeTags() {
     Collection<Tag> tags = new HashSet<>();
-    AwsUtils.appendResourceTags(tags, makeAwsCloudContext(), null);
+    AwsUtils.appendResourceTags(tags, AwsTestUtils.makeAwsCloudContext(), null);
     return tags;
-  }
-
-  // Cloud context
-
-  public static AwsCloudContext makeAwsCloudContext() {
-    return new AwsCloudContext(
-        new AwsCloudContextFields(
-            MAJOR_VERSION, ORGANIZATION_ID, ACCOUNT_ID, TENANT_ALIAS, ENVIRONMENT_ALIAS),
-        new CloudContextCommonFields(
-            DEFAULT_SPEND_PROFILE_ID, WsmResourceState.READY, /*flightId=*/ null, /*error=*/ null));
   }
 
   // S3 Folder
