+++ service/src/test/java/bio/terra/workspace/common/utils/WorkspaceUnitTestUtils.java
@@ -1,5 +1,7 @@
 package bio.terra.workspace.common.utils;
 
+import static bio.terra.workspace.common.fixtures.WorkspaceFixtures.DEFAULT_SPEND_PROFILE_ID;
+
 import bio.terra.workspace.common.fixtures.WorkspaceFixtures;
 import bio.terra.workspace.db.WorkspaceDao;
 import bio.terra.workspace.db.model.DbCloudContext;
@@ -18,7 +20,6 @@ import java.util.UUID;
 
 /** Utilities for working with workspaces in unit tests. */
 public class WorkspaceUnitTestUtils {
-  public static final SpendProfileId SPEND_PROFILE_ID = new SpendProfileId("my-spend-profile");
   public static final String POLICY_OWNER = "policy-owner";
   public static final String POLICY_WRITER = "policy-writer";
   public static final String POLICY_READER = "policy-reader";
@@ -39,7 +40,7 @@ public class WorkspaceUnitTestUtils {
   public static DbCloudContext makeDbCloudContext(CloudPlatform cloudPlatform, String json) {
     return new DbCloudContext()
         .cloudPlatform(cloudPlatform)
-        .spendProfile(WorkspaceUnitTestUtils.SPEND_PROFILE_ID)
+        .spendProfile(DEFAULT_SPEND_PROFILE_ID)
         .contextJson(json)
         .state(WsmResourceState.READY)
         .flightId(null)
@@ -68,7 +69,7 @@ public class WorkspaceUnitTestUtils {
       WorkspaceDao workspaceDao, UUID workspaceUuid, String projectId) {
     String flightId = UUID.randomUUID().toString();
     workspaceDao.createCloudContextStart(
-        workspaceUuid, CloudPlatform.GCP, SPEND_PROFILE_ID, flightId);
+        workspaceUuid, CloudPlatform.GCP, DEFAULT_SPEND_PROFILE_ID, flightId);
     workspaceDao.createCloudContextSuccess(
         workspaceUuid,
         CloudPlatform.GCP,
@@ -76,7 +77,7 @@ public class WorkspaceUnitTestUtils {
                 new GcpCloudContextFields(
                     projectId, POLICY_OWNER, POLICY_WRITER, POLICY_READER, POLICY_APPLICATION),
                 new CloudContextCommonFields(
-                    SPEND_PROFILE_ID, WsmResourceState.CREATING, flightId, /*error=*/ null))
+                    DEFAULT_SPEND_PROFILE_ID, WsmResourceState.CREATING, flightId, /*error=*/ null))
             .serialize(),
         flightId);
   }
