+++ service/src/test/java/bio/terra/workspace/db/WorkspaceDaoTest.java
@@ -1,11 +1,11 @@
 package bio.terra.workspace.db;
 
+import static bio.terra.workspace.common.fixtures.WorkspaceFixtures.DEFAULT_SPEND_PROFILE_ID;
 import static bio.terra.workspace.common.fixtures.WorkspaceFixtures.DEFAULT_USER_EMAIL;
 import static bio.terra.workspace.common.utils.WorkspaceUnitTestUtils.POLICY_APPLICATION;
 import static bio.terra.workspace.common.utils.WorkspaceUnitTestUtils.POLICY_OWNER;
 import static bio.terra.workspace.common.utils.WorkspaceUnitTestUtils.POLICY_READER;
 import static bio.terra.workspace.common.utils.WorkspaceUnitTestUtils.POLICY_WRITER;
-import static bio.terra.workspace.common.utils.WorkspaceUnitTestUtils.SPEND_PROFILE_ID;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
@@ -321,7 +321,7 @@ class WorkspaceDaoTest extends BaseUnitTest {
       // Mismatched flight id
       String flightId = UUID.randomUUID().toString();
       workspaceDao.createCloudContextStart(
-          workspaceUuid, CloudPlatform.GCP, WorkspaceUnitTestUtils.SPEND_PROFILE_ID, flightId);
+          workspaceUuid, CloudPlatform.GCP, DEFAULT_SPEND_PROFILE_ID, flightId);
 
       String gcpContextString = makeCloudContext().serialize();
 
@@ -387,7 +387,7 @@ class WorkspaceDaoTest extends BaseUnitTest {
     void workspaceCreateErrorDeserializes() {
       var flightId = UUID.randomUUID().toString();
       workspaceDao.createCloudContextStart(
-          workspaceUuid, CloudPlatform.GCP, SPEND_PROFILE_ID, flightId);
+          workspaceUuid, CloudPlatform.GCP, DEFAULT_SPEND_PROFILE_ID, flightId);
       var exception = new FieldSizeExceededException("This is a random ErrorReportException");
       workspaceDao.createCloudContextFailure(
           workspaceUuid,
@@ -434,7 +434,7 @@ class WorkspaceDaoTest extends BaseUnitTest {
         new GcpCloudContextFields(
             PROJECT_ID, POLICY_OWNER, POLICY_WRITER, POLICY_READER, POLICY_APPLICATION),
         new CloudContextCommonFields(
-            SPEND_PROFILE_ID, WsmResourceState.READY, /*flightId=*/ null, /*error=*/ null));
+            DEFAULT_SPEND_PROFILE_ID, WsmResourceState.READY, /*flightId=*/ null, /*error=*/ null));
   }
 
   private void checkCloudContext(Optional<GcpCloudContext> optionalContext) {
